feat/chore/refactor/test: Add `Menu` button in `Nav`, etc.
In `package.json` we add `lucide-react` and `prop-types` (because lucide
depend it, I think it maybe a bug, so I report [it][lucide-issue]). we
also move `tailwindcss` from `devDependecies` to `dependencies`.

We add `.main-part` for the body's main-part. It can change width by
device and used in two place, so I put it as a single class.

We also add `Menu` button in `Nav` and change the test for it. And make
the dark mode toggle looks good in dark mode.

We change a lot of things. But we think it can put in a single commit
because to split those are troublesome.

[lucide-issue]: https://github.com/lucide-icons/lucide/issues/486