feat/refactor/docs: Make TWO darkmode toggle have ONE state.
We add `redux` to hold the global state. But it need another package named
`react-it`. I add it and I do think it is a bug. See [here][our reply].
I think if we use `vite` instead of `snowpack`, this bug will be
resolved.

And we put those states in redux store. It works very good. But we do not
update the test and we will update it next commit.

And we update the to-do tree in document. We also need to change the document
to make it better. We will change it soon.

[our reply]: https://github.com/withastro/snowpack/issues/3818#issuecomment-1021980059