test/fix: Add test for localStorage and make test pass, etc.
We use `@testing-library/react` to test. Because it have more methods that only
DOM element.

We also add test for localStorage.

By the way, we fixed the bug -- It cannot show the right mode but just normal
mode. We will add a test next commit.

We use a long time to fix the bug that the data changed after we test it -- and
then the test failed. >_< sad. I record it [here][my-answer].

[my-answer]: https://stackoverflow.com/a/70867004/13031497