chore: Use `Vite` insteal of `Snowpack`, etc.
We find that `Vite` looks better than `Snowpack`. So we change it.

After change, we find that Jest can not run well because before there has a
plugin to deal the `.jsx` file to get the `.js` file and put the `.js` to Jest.
We add `Typescript` to do it. We also think using ts maybe is a good idea. So
we change file `jest.config.js` and add a file `tsconfig.json`.

We also remove `public` folder -- looks like that `vite` hate it. And we put
the original file in other place. And we also change `tailwind.config.js` to
make tailwind know the new place. And also change some source file in
`front/src/` folder.

we also add `autoprefixer` to deal with `.css` file and change file
`postcss.config.js`.