feat: Add clickable dots to `Carousel`.
We can click the dots to change what data to show.

By the way, we let the breakpoint be the `Carousel`'s props. Breakpoint means
that when we change the device's width greater than it, It will change the way
to show data.

In the `Carousel.tsx`, we can find a varable named `useless`. That's because
tailwindcss will read the `*.tsx` files, and remove class not used. So I add
those classes which we may use in the `useless` variable, to tell tailwindcss
do not remove those classes. But ... it is ugly, isn't it? I am going to try
to remove those and use CSS-in-JS.

I also add two class in `index.css`: `border-first` and `border-second`. To set
the buttones' border.

And I add something in `tailwind.config.js` to make tailwindcss have more
classes.