feat/docs: Use `styled-components` rather than `tailwindcss` in `Carousel`.
We use `styled-components` in `Carousel` component.

`tailwindcss` has mode `JIT`, which means that `just in build time` but not
`just in runtime`. So we need use `CSS-in-JS`.

We find `styled-system`, we think we will use it next commit. It looks like
have more readable breakpoint and have short CSS attr name.

Also update document.