feat: Add theme-ui, and make `Dots` use theme-ui instead of styled-system.
We add theme-ui, and config it because it has its own jsx. See [vite-jsx][] for
more infomation. So we change `package.json`, `yarn.lock`,
`tsconfig.json` and add file `vite.config.js`.

To use `theme-ui`, we add `anotherTheme.ts` (there is already `theme.ts` for
styled-component, we will try to remove it soon). And we change
`index.jsx` to use the theme.

Then we make `Dots` use theme-ui and change `Carousel` (because it use
`Dots`).

[vite-jsx]: https://github.com/system-ui/theme-ui/issues/2115.