feat/chore: Add new `Button` component with theme-ui, and use it in `Carousel`.
We now have two button: first one is using `theme-ui`, and the second is using
`styled-system`.

To use the new one, we replace the old button in `Carousel` component with the
new one.

The dark mode... Well, we will fix it as soon as we can.

By the way, to make the new button, we change the `theme-ui`'s theme file to
have moew meaningfull color name.