feat: Remove `theme` prop and add other props for `Button` component, etc.
We add props to `Button`:
- filp. set if hover, it will filp background and text's color. Or it will
  darken those colors. (So we add `@theme-ui/color` to deal those. And change
  file `package.json` and `yarn.lock`).
- color. set text color and border color.
- bg. set background.

Then we update `Carousel.tsx`, `index.jsx` because those are using `Button` or
something else. >_<. Those code are just shit.

In file `theme.ts`, we also change colors to before (but with little change).
And remove colors' `darker` and `lighter` attr because those attr are useless.