feat/refactor/chore: Remove tailwindcss, postcss, headlessui, etc.
We remove `tailwindcss`, `postcss` and `headlessui`. And those config file.

We also add some new function to set color and something else in `theme.ts` and
then use it in some components.

We also split code from `Nav` in `DarkToggle` and `Menu`.

By the way, we remove importToCDN, which is used in commit d748aa1f8696. It is
just a shit.