feat: Add page `Info`, etc.
We add page `Info` to show infomation. its data is set in `config.tsx`.

Because we need let the `Header` component in `Info` page has y-padding, so we
let the `Header` can dealwith `className` prop.

To let `Info` page can return to index page, we let the `Nav` component have a
`Link` component to return, and in `index.jsx`, we set the global `a` element
style (because `Link` component will be render as `a` element).