feat: Add page `News`.
Well, we have `News` component, and now we also have `News` page. Well. Well.
We will going to rename it or something else.

To go to `News` page, we let the button `Read More` in component `Carousel` be
clickable. To let the news have its own path, we add `id` for those.

To mock the new `id` attr, we also change the mock file `handler.js`. and let
it can deal with the query with variable. (By the way, to make it have better
data, we are using `faker.js`, I really hope that Marak get the money he should
have. Damn it, fuck you Github).

After that, we add its page add its `Router` in `index.jsx`. After that I find
that we need to scroll to top auto. So I add component `ScrollToTop` to deal
with it.

OK, that's all.