feat/refactor/style: Make the `DarkToggle` have more props, etc.
Now `DarkToggle` have `size` prop, it has the same function as `Button`. It can
set its size now, and it is easy to let it have the same size with `Button`.

And by the way, `DarkToggle` have `showText` prop. It can set to show the text
`Dark Mode` or not.

Finally, I make also element's default font-family to be `inherit`. And I
cahange some function name to have better meaning, and change some code's
style.