feat: Using function to get HSV color value.
Before we have a lot of colors that just a value. Now we using a function to
build those colors. This way will reduce the file's size and make the code
more clear and easy to understand.

Add `outline` part in `theme.ts`.

By the way, we add some key-value pair and change some key's value in file
`theme.ts`to make it looks better.

`@theme-ui/color` can not deal with HSV. So we remove it.

We will remove `flip` prop in `Button`. And we will add `hover` argument
for `setXxxxxx` function in `theme.ts`.