feat/docs: Add sign out button.
(But it can't be clicked now, we will try to put this data into redux to
handle.)

We add mock: currentUser to handle the GraphQL query `CurrentUser`. It can tell
us if we are logged in.

By the way, we renameed the file `Sign.ts` to `Auth.ts`. And let the handler
files export an array of functions, not just return only one function. (Because
now one file will export many functions instead of previous only one function)

By the way, we renameed the file `Sign.ts` to `Auth.ts`. And let the handler
files export an array of functions, not just return only one function. (Because
now one file will export many functions, instead of only exporting one function
before)