docs: Add register page, etc.
First at all, I rename page `Sign.tsx` into `SignIn.tsx`. Because this page
only handle `sign in` but not `sign up`. I put the `sign up` function on page
`Register.tsx`.

Set URL in page `SignIn.tsx`, to avoid having multipie string in different
source files. So I change component `SignButton`, and the entrypoint
`index.jsx`.

Put the component `Input` outside the page `SignIn`, because it is used not
only in page `SignIn` but also in page `Register`. And I also put the boxSx,
which is a style object to show the page `SignIn`'s sign in outside box, in to
`config.tsx`.

Also update the documents.