feat/style/refactor: Add hook to handler sign up event.
In file `authSlice.ts`, add basic function to send mutation `SignUp` and
put those in a hook `useSignUp` (We should change the authentication's state in
the process of sending mutation in the future). We also add some comment to
make the code more readable.

In page `Register`, I add a input box for `name`, and call the function
`signUp` (from the hook `useSignUp`) if user click the button `resgister`.

In mock handlers, we add a handler for mutation `SignUp` (in file `Auth.ts`).