feat/style(back): Add `isAdmin` field, etc, add signIn.
We need to make sure if the user is admin. So we add `isAdmin` field.

We also:
- Add mutation `signIn`.
- Rename mutation `createUser` into `register`.

Add we format some SQL statements.

To make sure that `createUser` can set the cookie in the GraphQL resolve
function, we put the `req` and `res` in those functions' context (because we
need method `res.cookie` to set the response's `Set-Cookie`).