fix/style(back): Make the backend can hold large image.
We use the data URL to send image in GraphQL. But it failed because the data
URL is TOO LARGE for MySQL/MariaDB and the GraphQL server.

So I set the type of `image_url` in table `news` to be `MEDIUMTEXT`. Now it can
hold data URLs smaller than 16MB.

And I also set the server's json size limit to 20MB.

We also change the style of some code, and rename the `MutationType`(a GraphQL
object type)'s name from `createNews`(not at all) to `mutation`.