Chore: Make full project as a monorepo (yarn workspace)
More information about yarn workspace:

> https://classic.yarnpkg.com/en/docs/workspaces/

Why we choose monorepo (yarn workspace): monorepo is a way to hold the project,
which contains kinds of packages. Now we already have two packages, back-end
and front-end, and that's why we need monorepo to control the two packages.

If there are some scripts for all packages (for example, a script that using
prettier to format code), it will be a better idea that put it at the root
workspace rather than put the same script to two or more package.