Chore(front): Make `front` can use `theme-shortcut`
We put the path of the package `@acm-homepage/theme-shortcut` into the
`workspaces` array, which is in the file `package.json` that is at the root of
project. So that the package is a part of the project now.

We also let the package `@acm-homepage/theme-shortcut` be the dependency of the
package `@acm-homepage/front` by changing the file `front/packages.json`. So
that the code, which is in the package `@acm-homepage/front`, can use anything
that is exported by the package `@acm-homepage/theme-shortcut` now.