Refactor(front): Use `theme-shortcut` in `front`
package `theme-shortcut` is now used in folder `front/src/component/`, folder
`front/src/page/` and file `front/util/theme.ts` that is in pacakge `front`,
and we try our best to reduce the dependence on file `front/src/util/theme.ts`.

The style objects, which are in package `front`, are more clear than before
after we use the package `theme-shortcut`.