Feat(front): Add more input components
Now we add more input components in the folder `front/src/component/InputSet`:

* `ContextTextArea`: The textarea to edit context.
* `EmailInput`: The input box to edit email.
* `ImageInput`: The image uploader to upload image.
* `NicknameInput`: The nickname input box to edit nickname.
* `PasswordInput`: The input box to edit password.
* `TitleInput`: The input box to edit title.

The components are built on component `Input`. The component `Input` is changed
to only provide the container and Theme-UI's style object.

We update pages to use those components in `InputSet` rather than original
component `Input` as well.