refactor(back): Replacing the back with Typescript
In this commit, I am trying to refactoring the back with Typescript very carefull
y.

In detail, I renamed `main.js`, `salt.js` to `main.ts` and `salt.ts`. To suppor
t this change, I created a file named `tsconfig.ts` at the root of the back to
declare that this project is a Typescript project. So we are able to compile th
is by `cd ./back & tsc`. This will generate a dir named `build` where there are
the compiled Javascript files.

Since we have generated the Js files, we should revise the `package.json` file
to update commands to run the Js files. Check `package.json` to get more.

Through my test, I found some import statements is not available in some files.
So I revised them to pass the compilation.

Finally, the `.gitignore` file is updated to filter `./back/build/*`.