feat(front): Update the component `News`.
Now we can get the `News` in folder `front/src/component/News/News.tsx` now.

I put all sub-components of this component `News` into the same folder. To make
the struct of the project more clear.

And update `Index`, which is a page, and use the old component `News` before.