feat(front): Update the component `Nav`.
Now we can get the `Nav` in folder `front/src/component/Nav/Nav.tsx` now.

I put all sub-components of this component `Nav` into the same folder. To make
the struct of the project more clear.

And update `Index`, which is a page, and use the old component `Nav` before.