feat(back): Complete the back-end interface for crawler data storage
In this commit, I made the back-end enable to store the data from the crawler.

To extend, I created 2 new tables named `userAccount` and `problemSolved`.
`UserAccount` refers to the user-account relation and `problemSolved` intends
to store the contest data for the account.

I also made the back-end support `getUserAccount`, `getSovled`,
`insertUserAccount`, `removeUserAccount`, `insertSolved` and `removeSolved`!